move ensure index to spandex bootstrap Revert "use parens to show a method with side effect" Revert "move ensure index to bootstrap"
This reverts commit 57d856a0b78eecedafb02e054e73d1136bbae30f.
This reverts commit 23494fbcf51c2e6a72103b7c495c28af74c3c74b.
