+++ src/main/scala/com/socrata/spandex/Spandex.scala
@@ -9,6 +9,8 @@ object Spandex extends App {
   val conf = new SpandexConfig
 
   override def main(args: Array[String]): Unit = {
+    SpandexBootstrap.ensureIndex(conf)
+
     val port = conf.port
 
     val context = new WebAppContext
